+++ src/Network/Network/ChangeLog.md
@@ -31,6 +31,8 @@
 * Updated cmdlets to enable setting of ConnectionMode on Virtual Network Gateway Connections.
     - `New-AzVirtualNetworkGatewayConnection`
     - `Set-AzVirtualNetworkGatewayConnection`
+* Updated `New-AzFirewallPolicyApplicationRule` cmdlet:
+    - Added parameter WebCategory
 
 ## Version 4.3.0
 * Updated below cmdlet